requirements.txt
tensorflow>=2.9.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
opencv-python>=4.5.0
tqdm>=4.60.0
scikit-learn>=0.24.0
pyyaml>=6.0
wandb>=0.12.0
tensorflow-graphics>=1.0.0
